requirements.txt
gspread
google-auth
openai
openpyxl
python-dotenv
